# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 611b3b87fb1f2d7735e87134
  title: wlg
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /gps:
    post:
      x-skaffolder-id: 611b41defb1f2d7735e871e5
      x-skaffolder-name: create
      x-skaffolder-id-resource: 611b4155fb1f2d7735e871cc
      x-skaffolder-resource: GPS
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - GPS
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GPS"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Imei:
                  type: number
                Models:
                  type: string
                SN:
                  type: number
    get:
      x-skaffolder-id: 611b41defb1f2d7735e871ea
      x-skaffolder-name: list
      x-skaffolder-id-resource: 611b4155fb1f2d7735e871cc
      x-skaffolder-resource: GPS
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - GPS
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GPS"
  "/gps/{id}":
    delete:
      x-skaffolder-id: 611b41defb1f2d7735e871e9
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 611b4155fb1f2d7735e871cc
      x-skaffolder-resource: GPS
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - GPS
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 611b41defb1f2d7735e871e1
      x-skaffolder-name: get
      x-skaffolder-id-resource: 611b4155fb1f2d7735e871cc
      x-skaffolder-resource: GPS
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - GPS
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GPS"
    post:
      x-skaffolder-id: 611b41defb1f2d7735e871e3
      x-skaffolder-name: update
      x-skaffolder-id-resource: 611b4155fb1f2d7735e871cc
      x-skaffolder-resource: GPS
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - GPS
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GPS"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GPS"
  /sim:
    post:
      x-skaffolder-id: 611b41d1fb1f2d7735e871d7
      x-skaffolder-name: create
      x-skaffolder-id-resource: 611b4133fb1f2d7735e871c5
      x-skaffolder-resource: Sim
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Sim
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sim"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                SimNo:
                  type: number
    get:
      x-skaffolder-id: 611b41d2fb1f2d7735e871dd
      x-skaffolder-name: list
      x-skaffolder-id-resource: 611b4133fb1f2d7735e871c5
      x-skaffolder-resource: Sim
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Sim
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sim"
  "/sim/{id}":
    delete:
      x-skaffolder-id: 611b41d2fb1f2d7735e871de
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 611b4133fb1f2d7735e871c5
      x-skaffolder-resource: Sim
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Sim
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 611b41d1fb1f2d7735e871d6
      x-skaffolder-name: get
      x-skaffolder-id-resource: 611b4133fb1f2d7735e871c5
      x-skaffolder-resource: Sim
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Sim
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sim"
    post:
      x-skaffolder-id: 611b41d1fb1f2d7735e871d5
      x-skaffolder-name: update
      x-skaffolder-id-resource: 611b4133fb1f2d7735e871c5
      x-skaffolder-resource: Sim
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Sim
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sim"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sim"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GPS:
      x-skaffolder-id: 611b4155fb1f2d7735e871cc
      x-skaffolder-id-db: 611b3b87fb1f2d7735e87135
      x-skaffolder-id-entity: 611b4155fb1f2d7735e871c8
      x-skaffolder-url: /gps
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Imei:
          type: number
          x-skaffolder-id-attr: 611b4155fb1f2d7735e871c9
          x-skaffolder-type: Number
          x-skaffolder-required: true
        Models:
          type: string
          x-skaffolder-id-attr: 611b4155fb1f2d7735e871cb
          x-skaffolder-type: String
        SN:
          type: number
          x-skaffolder-id-attr: 611b4155fb1f2d7735e871ca
          x-skaffolder-type: Number
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - Imei
        - SN
    Sim:
      x-skaffolder-id: 611b4133fb1f2d7735e871c5
      x-skaffolder-id-db: 611b3b87fb1f2d7735e87135
      x-skaffolder-id-entity: 611b4133fb1f2d7735e871c3
      x-skaffolder-url: /sim
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        SimNo:
          type: number
          x-skaffolder-id-attr: 611b4133fb1f2d7735e871c4
          x-skaffolder-type: Number
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - SimNo
...